Implement rendering-independent physics simulation
This seems to improve the stability of the simulation between runs by
quite a lot. Now the only time the rendering affects the simulation is
if it takes way too long (i.e. the app goes out of focus), since when
that happens we pause the simulation so we don't end up spiraling out of
control trying to catchup with the accumulator.